fix: get_ranking RPC 리턴 타입 INTEGER로 수정 (scores 테이블과 일치)

+++ supabase_migration.sql
@@ -32,8 +32,8 @@ CREATE OR REPLACE FUNCTION get_ranking(
 )
 RETURNS TABLE (
   rank BIGINT,
-  score BIGINT,
-  height BIGINT,
+  score INTEGER,
+  height INTEGER,
   nickname TEXT,
   country_code TEXT,
   user_id UUID,
